+++ http/java/okhttp/lib/src/main/java/com/microsoft/kiota/http/middleware/options/TelemetryHandlerOption.java
@@ -1,0 +1,23 @@
+package com.microsoft.kiota.http.middleware.options;
+
+import java.util.function.Function;
+
+import com.microsoft.kiota.RequestOption;
+
+import okhttp3.Request;
+
+/**
+ * TelemetryHandlerOption class
+ */
+public class TelemetryHandlerOption implements RequestOption {
+
+    /**
+    * A delegate which can be called to configure the Request with desired telemetry values.
+    */
+    public Function<Request, Request> TelemetryConfigurator = (request) -> request;
+
+    @Override
+    public <T extends RequestOption> Class<T> getType() {
+        return (Class<T>) TelemetryHandlerOption.class; 
+    }
+}